Added getHumanChoice()
This function prompts the user to enter a number
for one of the following choices: 1. Rock,
2. Paper, or 3. Scissors. If the user enters
something other than an integer, the script
outputs an alert message to let them know that
they must enter an integer between 1 and 3. The
function then returns the value in the array
`strings` associated with the index userInput-1.