Added playRound(humanChoice, computerChoice)
This function lets you play a round of Rock Paper
Scissors with the computer. The constants
humanChoice and computerChoice call the
getHumanChoice() and getComputerChoice() functions
respectively. These constants are then passed as
arguments to a playRound() function call. Within
the function, values of humanChoice and
computerChoice are checked against each other to
determine a winner or if there was a tie.