Changes to index.html and script.js
In index.html: added an h1 heading, a description
of the rules of rock paper scissors, and a button
to start the game. The button's onclick property
will call the function playGame(). This function
allows the user to play the game for 5 rounds by
dynamically calling the playRound() function nested
within it. The humanScore and computerScore
variables are incremented each round depending on
whoever's the winner or not. After the five rounds,
the final winner of the game is calculated and
shown to the user.