Cleaned up code and logic in script.js
Debugged an error on line 84 where
querySelector("#results") was returning null.
The problem was that the query was not specific
enough, so the string "body > #results" had to be
passed to the function.

Also fixed another logic error on line 91 where
event listeners are added to each choice
button. No runtime error was raised, but the problem
was that the playRound() function was being passed
as an argument to the event listener instead of being
a callback function.